requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
distro==1.9.0
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
openai==1.31.1
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-file==2.1.0
sniffio==1.3.1
tldextract==5.1.2
tqdm==4.66.4
typing_extensions==4.12.1
urllib3==2.2.1
Werkzeug==3.0.3
yarl==1.9.4
zope.event==5.0
zope.interface==6.4.post2